Refactor smoke-test setup
