+++ smoke-tests/app-template/app/router.js
@@ -7,5 +7,4 @@ export default class Router extends EmberRouter {
 }
 
 Router.map(function () {
-  this.route('example-gjs-route')
 });
