--- smoke-tests/app-template/app/routes/example-gjs-route.js
@@ -1,8 +1,0 @@
-import Route from '@ember/routing/route';
-export default class extends Route {
-  model() {
-    return {
-      message: "I am the model"
-    }
-  }
-}
